Refactor: reafcotr related to restaurants route & add error function on helpers
