+++ routes/comment.js
@@ -2,13 +2,7 @@ const express = require('express')
 const router = express.Router()
 const auth = require('../middleware/auth')
 const commentController = require('../controllers/commentController')
-const handleErrorAsync = func => async (req, res, next) => {
-  try {
-    await func(req, res, next)
-  } catch (error) {
-    next(error)
-  }
-}
+const handleErrorAsync = require('../_helpers').handleErrorAsync
 router.use(auth.authenticated)
 
 router.post('/', handleErrorAsync(async (req, res, next) => {
