+++ routes/home.js
@@ -3,41 +3,9 @@ const router = express.Router()
 const userController = require('../controllers/userController')
 const passport = require('passport')
 const auth = require('../middleware/auth')
-const restController = require('../controllers/restController')
-const handleErrorAsync = func => async (req, res, next) => {
-  try {
-    await func(req, res, next)
-  } catch (error) {
-    next(error)
-  }
-}
+const handleErrorAsync = require('../_helpers').handleErrorAsync
 
 router.get('/', auth.authenticated, (req, res) => res.redirect('/restaurants'))
-
-//顯示所有餐廳頁面
-router.get('/restaurants', auth.authenticated, handleErrorAsync(async (req, res, next) => {
-  const { data, categories, categoryId, totalPage, prev, nextPage, page } = await restController.getRestaurants(req.query)
-  return res.render('restaurants', { restaurants: data, categories, categoryId, totalPage, prev, nextPage, page })
-}))
-
-//Feeds
-router.get('/restaurants/feeds', handleErrorAsync(async (req, res, next) => {
-  const { restaurants, comments } = await restController.getFeeds()
-  res.render('feeds', { restaurants, comments })
-}))
-//dashboard
-router.get('/restaurant/dashboard/:id', handleErrorAsync(async (req, res, next) => {
-  const restaurant = await restController.getRestaurant(req.params.id)
-  const totalCount = await restController.getTotalCountOfComment(req.params.id)
-  return res.render('dashboard', { restaurant, totalCount })
-}))
-
-//顯示單一餐廳頁面
-router.get('/restaurants/:id', handleErrorAsync(async (req, res, next) => {
-  const restaurant = await restController.getRestaurant(req.params.id)
-  await restController.calculatorViewCounts(req.params.id)
-  return res.render('restaurant', { restaurant })
-}))
 
 router.get('/signup', (req, res) => {
   return res.render('signup')
