+++ views/restaurant.handlebars
@@ -15,7 +15,7 @@
   </div>
   <div class="col-md-8">
     <p>{{restaurant.description}}</p>
-    <a href="/restaurant/dashboard/{{restaurant.id}}" type="button" class="btn btn-primary">DashBoard</a>
+    <a href="/restaurants/dashboard/{{restaurant.id}}" type="button" class="btn btn-primary">DashBoard</a>
   </div>
   <hr>
 </div>
